Fix so that signed integers use two's compliment
We start of with a 64-bit unsigned integer that is max value.
Then we shift away the size of the signal. Basically setting zeroes to from pos 0-signal lenght.
Then we add the signal we have by using or on all the values that are now zero.
In short we pad the signal with one so that we can take out a signed 64-bit integer