feat: Move handling of send result to function
This is done so as to avoid code duplication and simplification in
general.

This also includes an important change related to retries. Until now,
retries were attempted with an increasing "back-off" sleep
time. Hitherto, the time to sleep between the attempts is randomized
each time in an interval between the min and max value as declared in
the configuration file. It is therefore important to review these
values to achieve the intended result.