feat: Use numbers for digital values instead of bools
It is more common to use 1 or 0 to represent the states of a digital
port than true or false. The Rust GPIO library uses u8 which does not
exist in protobuf. In the latter case, therefore, i32 is used instead.

This is a breaking change for all users of digital I/O since
true/false must be replaced by 1/0.