feat: Rework software upgrades
If the current major version is greater than or equal to the required
version, upgrade the package to the latest available minor + patch
version.

If the required major version is higher than the current version,
uninstall the current version and install the new one.

If the new major version could not be installed, reinstall the old one.